Fixed comments from review3
