+++ Modsen.Server.Authentication.Api/Controllers/AuthenticationController.cs
@@ -4,7 +4,6 @@ using Modsen.Server.Authentication.Application.Helpers;
 using Modsen.Server.Authentication.Application.Models.Authentication;
 using MediatR;
 using Modsen.Server.Authentication.Application.Features.ApplicationUser.Commands;
-using Microsoft.AspNetCore.Authorization;
 
 namespace Modsen.Server.Authentication.Api.Controllers
 {
@@ -21,53 +20,30 @@ namespace Modsen.Server.Authentication.Api.Controllers
         [Route("register")]
         public async Task<IActionResult> Register(RegisterModel registerModel)
         {
-            try
+            var tokenModel = await _mediator.Send(new RegisterApplicationUser
             {
-                var tokenModel = await _mediator.Send(new RegisterApplicationUser
-                {
-                    RegisterModel = registerModel,
-                    RefreshTokenValidityInDays = _refreshTokenValidityInDays
-                });
+                RegisterModel = registerModel,
+                RefreshTokenValidityInDays = _refreshTokenValidityInDays
+            });
 
-                CookieHelper.SetRefreshTokenInCookie(tokenModel.RefreshToken, _refreshTokenValidityInDays, HttpContext);
+            CookieHelper.SetRefreshTokenInCookie(tokenModel.RefreshToken, _refreshTokenValidityInDays, HttpContext);
 
-                return Ok(new
-                {
-                    accessToken = tokenModel.AccessToken
-                });
-            }
-            catch (Exception ex)
-            {
-                return StatusCode(403, new
-                {
-                    errors = ex.Message
-                });
-            }
+            return Ok(tokenModel.AccessToken);
         }
 
         [HttpPost]
         [Route("login")]
         public async Task<IActionResult> Login(LoginModel loginModel)
         {
-            try
+            var tokenModel = await _mediator.Send(new LoginApplicationUser
             {
-                var tokenModel = await _mediator.Send(new LoginApplicationUser
-                {
-                    LoginModel = loginModel,
-                    RefreshTokenValidityInDays = _refreshTokenValidityInDays
-                });
+                LoginModel = loginModel,
+                RefreshTokenValidityInDays = _refreshTokenValidityInDays
+            });
 
-                CookieHelper.SetRefreshTokenInCookie(tokenModel.RefreshToken, _refreshTokenValidityInDays, HttpContext);
+            CookieHelper.SetRefreshTokenInCookie(tokenModel.RefreshToken, _refreshTokenValidityInDays, HttpContext);
 
-                return Ok(new
-                {
-                    accessToken = tokenModel.AccessToken
-                });
-            }
-            catch
-            {
-                return Unauthorized();
-            }
+            return Ok(tokenModel.AccessToken);
         }
     }
 }
