--- Modsen.Server.Authentication.Api/DependencyInjection.cs
@@ -1,59 +1,0 @@
-﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
-using Modsen.Server.Authentication.Domain.Entities;
-using Modsen.Server.Authentication.Domain.Interfaces.Services;
-using Modsen.Server.Authentication.Infrastructure.Data;
-using Modsen.Server.Authentication.Infrastructure.Services;
-using System.Text;
-
-namespace Modsen.Server.Authentication.Api
-{
-    public static class DependencyInjection
-    {
-        public static IServiceCollection AddPresentation(this IServiceCollection services, WebApplicationBuilder builder)
-        {
-            services.AddDbContext<DBContext>(options =>
-                options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
-
-            services.AddIdentity<ApplicationUser, IdentityRole>(options =>
-            {
-                options.Password.RequireDigit = false;
-                options.Password.RequiredLength = 1;
-                options.Password.RequireLowercase = false;
-                options.Password.RequireUppercase = false;
-                options.Password.RequireNonAlphanumeric = false;
-            })
-            .AddEntityFrameworkStores<DBContext>()
-            .AddDefaultTokenProviders();
-
-            services
-                .AddAuthentication(options =>
-                {
-                    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-                    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-                    options.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
-                })
-                .AddJwtBearer(options =>
-                {
-                    options.SaveToken = true;
-                    options.RequireHttpsMetadata = false;
-                    options.TokenValidationParameters = new TokenValidationParameters()
-                    {
-                        ValidateIssuer = true,
-                        ValidateAudience = true,
-                        ValidateLifetime = true,
-                        ValidateIssuerSigningKey = true,
-                        ValidAudience = builder.Configuration["JWT:ValidAudience"],
-                        ValidIssuer = builder.Configuration["JWT:ValidIssuer"],
-                        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(builder.Configuration["JWT:SecretKey"]!)),
-                    };
-                });            
-
-            services.AddScoped<ITokenProviderService, TokenProviderService>();
-
-            return services;
-        }
-    }
-}
