+++ Modsen.Server.Authentication.Api/Helpers/AuthenticateHelper.cs
@@ -16,14 +16,14 @@ namespace Modsen.Server.Authentication.Api.Helpers
 
         public static string GetAccessToken(HttpContext httpContext)
         {
-            var authorizationValue = httpContext.Request.Headers.Authorization[0] ?? throw new BadRequestException();
+            var authorizationValue = httpContext.Request.Headers.Authorization[0] ?? throw new ArgumentException();
 
             return authorizationValue.Split(' ')[1];
         }
 
         public static string GetRefreshToken(HttpContext httpContext)
         {
-            return httpContext.Request.Cookies["RefreshToken"] ?? throw new BadRequestException();
+            return httpContext.Request.Cookies["RefreshToken"] ?? throw new ArgumentException();
         }
     }
 }
