+++ Modsen.Server.Authentication.Api/MiddlewareExtensions/UseExceptionHandlerMiddlewareExtension.cs
@@ -1,0 +1,12 @@
+﻿using Modsen.Server.Authentication.Api.Middlewares;
+
+namespace Modsen.Server.Authentication.Api.MiddlewareExtensions
+{
+    public static class UseExceptionHandlerMiddlewareExtension
+    {
+        public static void UseExceptionHandlerMiddleware(this IApplicationBuilder app)
+        {
+            app.UseMiddleware<ExceptionHandlerMiddleware>();
+        }
+    }
+}
