+++ Modsen.Server.Authentication.Application/DependencyInjection.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using Modsen.Server.Authentication.Application.Features.ApplicationUser.Queries;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Commands;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Queries;
 
 namespace Modsen.Server.Authentication.Application
 {
@@ -11,17 +8,6 @@ namespace Modsen.Server.Authentication.Application
         {
             services.AddMediatR(configuration =>
                 configuration.RegisterServicesFromAssembly(typeof(DependencyInjection).Assembly));
-
-            services.AddTransient<AddApplicationUserToRoleByIdUseCase>();
-            services.AddTransient<ChangeApplicationUserRefreshTokenUseCase>();
-            services.AddTransient<ChangeApplicationUserStatusHandlerUseCase>();
-            services.AddTransient<DenyApplicationUserRoleByIdUseCase>();
-            services.AddTransient<RegisterUserUseCase>();
-            services.AddTransient<LoginUserUseCase>();
-            services.AddTransient<RefreshTokenUseCase>();
-
-            services.AddTransient<GetApplicationUserByUsernameUseCases>();
-            services.AddTransient<GetApplicationUsersUseCase>();
 
             return services;
         }
