+++ Modsen.Server.Authentication.Application/Features/ApplicationUser/CommandHandlers/ChangeApplicationUserRefreshTokenHandler.cs
@@ -2,7 +2,6 @@
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using Modsen.Server.Authentication.Application.Features.ApplicationUser.Commands;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Commands;
 using Modsen.Server.Authentication.Domain.Exeptions;
 using System;
 using System.Collections.Generic;
@@ -13,17 +12,19 @@ using System.Threading.Tasks;
 
 namespace Modsen.Server.Authentication.Application.Features.ApplicationUser.CommandHandlers
 {
-    public class ChangeApplicationUserRefreshTokenHandler(ChangeApplicationUserRefreshTokenUseCase changeApplicationUserRefreshTokenUseCase) 
+    public class ChangeApplicationUserRefreshTokenHandler(UserManager<Domain.Entities.ApplicationUser> userManager)
         : IRequestHandler<ChangeApplicationUserRefreshToken>
     {
-        private readonly ChangeApplicationUserRefreshTokenUseCase _changeApplicationUserRefreshTokenUseCase = changeApplicationUserRefreshTokenUseCase;
+        private readonly UserManager<Domain.Entities.ApplicationUser> _userManager = userManager;
 
         public async Task Handle(ChangeApplicationUserRefreshToken request, CancellationToken cancellationToken)
         {
-            await _changeApplicationUserRefreshTokenUseCase.ChangeRefreshTokenAsync(
-                request.UserName,
-                request.RefreshToken,
-                request.RefreshTokenValidityInDays);
+            var user = await _userManager.FindByNameAsync(request.UserName) ?? throw new NotFoundException();
+
+            user.RefreshToken = request.RefreshToken;
+            user.RefreshTokenExpiryTime = DateTime.Now.AddDays(request.RefreshTokenValidityInDays);
+
+            await _userManager.UpdateAsync(user);
         }
     }
 }
