+++ Modsen.Server.Authentication.Application/Features/ApplicationUser/CommandHandlers/ChangeApplicationUserStatusHandler.cs
@@ -1,22 +1,22 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Modsen.Server.Authentication.Application.Features.ApplicationUser.Commands;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Commands;
 using Modsen.Server.Authentication.Domain.Exeptions;
 
 namespace Modsen.Server.Authentication.Application.Features.ApplicationUser.CommandHandlers
 {
-    public class ChangeApplicationUserStatusHandler(ChangeApplicationUserStatusHandlerUseCase changeApplicationUserStatusHandlerUseCase)
+    public class ChangeApplicationUserStatusHandler(UserManager<Domain.Entities.ApplicationUser> userManager)
         : IRequestHandler<ChangeApplicationUserStatus>
     {
-        private readonly ChangeApplicationUserStatusHandlerUseCase _changeApplicationUserStatusHandlerUseCase 
-            = changeApplicationUserStatusHandlerUseCase;
+        private readonly UserManager<Domain.Entities.ApplicationUser> _userManager = userManager;
 
         public async Task Handle(ChangeApplicationUserStatus request, CancellationToken cancellationToken)
         {
-            await _changeApplicationUserStatusHandlerUseCase.ChangeUserStatusAsync(
-                request.UserId,
-                request.Status);
+            var user = await _userManager.FindByIdAsync(request.UserId) ?? throw new NotFoundException();
+
+            user.UserStatus = request.Status;
+
+            await _userManager.UpdateAsync(user);
         }
     }
 }
