+++ Modsen.Server.Authentication.Application/Features/ApplicationUser/CommandHandlers/DenyApplicationUserRoleByIdHandler.cs
@@ -1,20 +1,20 @@
 ﻿using MediatR;
+using Microsoft.AspNetCore.Identity;
 using Modsen.Server.Authentication.Application.Features.ApplicationUser.Commands;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Commands;
+using Modsen.Server.Authentication.Domain.Exeptions;
 
 namespace Modsen.Server.Authentication.Application.Features.ApplicationUser.CommandHandlers
 {
-    public class DenyApplicationUserRoleByIdHandler(DenyApplicationUserRoleByIdUseCase denyApplicationUserRoleByIdUseCase)
+    public class DenyApplicationUserRoleByIdHandler(UserManager<Domain.Entities.ApplicationUser> userManager)
         : IRequestHandler<DenyApplicationUserRoleById>
     {
-        private readonly DenyApplicationUserRoleByIdUseCase _denyApplicationUserRoleByIdUseCase 
-            = denyApplicationUserRoleByIdUseCase;
+        private readonly UserManager<Domain.Entities.ApplicationUser> _userManager = userManager;
 
         public async Task Handle(DenyApplicationUserRoleById request, CancellationToken cancellationToken)
         {
-            await _denyApplicationUserRoleByIdUseCase.DenyRoleAsync(
-                request.UserId,
-                request.RoleName);
+            var user = await _userManager.FindByIdAsync(request.UserId) ?? throw new NotFoundException();
+
+            await _userManager.RemoveFromRoleAsync(user, request.RoleName);
         }
     }
 }
