+++ Modsen.Server.Authentication.Application/Features/ApplicationUser/QueryHandlers/GetApplicationUserByUsernameHandler.cs
@@ -1,21 +1,18 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Modsen.Server.Authentication.Application.Features.ApplicationUser.Queries;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Queries;
 using Modsen.Server.Authentication.Domain.Exeptions;
 
 namespace Modsen.Server.Authentication.Application.Features.ApplicationUser.QueryHandlers
 {
-    public class GetApplicationUserByUsernameHandler(GetApplicationUserByUsernameUseCases getApplicationUserByUsernameUseCases)
+    public class GetApplicationUserByUsernameHandler(UserManager<Domain.Entities.ApplicationUser> userManager)
         : IRequestHandler<GetApplicationUserByUsername, Domain.Entities.ApplicationUser>
     {
-        private readonly GetApplicationUserByUsernameUseCases _getApplicationUserByUsernameUseCases 
-            = getApplicationUserByUsernameUseCases;
+        private readonly UserManager<Domain.Entities.ApplicationUser> _userManager = userManager;
 
         public async Task<Domain.Entities.ApplicationUser> Handle(GetApplicationUserByUsername request, CancellationToken cancellationToken)
         {
-            return await _getApplicationUserByUsernameUseCases.GetAsync(
-                request.Username);
+            return await _userManager.FindByNameAsync(request.Username) ?? throw new NotFoundException();
         }
     }
 }
