+++ Modsen.Server.Authentication.Application/Features/ApplicationUser/QueryHandlers/GetApplicationUsersHandler.cs
@@ -2,21 +2,20 @@
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Modsen.Server.Authentication.Application.Features.ApplicationUser.Queries;
-using Modsen.Server.Authentication.Application.UseCases.ApplicationUser.Queries;
 
 namespace Modsen.Server.Authentication.Application.Features.ApplicationUser.QueryHandlers
 {
-    public class GetApplicationUsersHandler(GetApplicationUsersUseCase getApplicationUsersUseCase)
+    public class GetApplicationUsersHandler(UserManager<Domain.Entities.ApplicationUser> userManager)
         : IRequestHandler<GetApplicationUsers, List<Domain.Entities.ApplicationUser>>
     {
-        private readonly GetApplicationUsersUseCase _getApplicationUsersUseCase = getApplicationUsersUseCase;
+        private readonly UserManager<Domain.Entities.ApplicationUser> _userManager = userManager;
 
         public async Task<List<Domain.Entities.ApplicationUser>> Handle(GetApplicationUsers request, CancellationToken cancellationToken)
         {
-            return await _getApplicationUsersUseCase.GetUsersAsync(
-                request.Page,
-                request.Count,
-                cancellationToken);
+            return await _userManager.Users
+                .Skip(request.Page * request.Count)
+                .Take(request.Count)
+                .ToListAsync(cancellationToken);
         }
     }
 }
